Migrated tables to new format and removed test tables
