+++ azuread/table_azuread_conditional_access_policy.go
@@ -2,14 +2,19 @@ package azuread
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"strings"
 
-	"github.com/manicminer/hamilton/msgraph"
-	"github.com/manicminer/hamilton/odata"
+	"github.com/iancoleman/strcase"
 	"github.com/turbot/steampipe-plugin-sdk/v3/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v3/plugin/transform"
 
 	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+
+	msgraphcore "github.com/microsoftgraph/msgraph-sdk-go-core"
+	"github.com/microsoftgraph/msgraph-sdk-go/identity/conditionalaccess/policies"
+	"github.com/microsoftgraph/msgraph-sdk-go/models"
 )
 
 //// TABLE DEFINITION
@@ -19,13 +24,17 @@ func tableAzureAdConditionalAccessPolicy() *plugin.Table {
 		Name:        "azuread_conditional_access_policy",
 		Description: "Represents an Azure Active Directory (Azure AD) Conditional Access Policy",
 		Get: &plugin.GetConfig{
-			Hydrate:           getAdConditionalAccessPolicy,
-			ShouldIgnoreError: isNotFoundErrorPredicate([]string{"Invalid object identifier"}),
-			KeyColumns:        plugin.SingleColumn("id"),
+			Hydrate: getAdConditionalAccessPolicy,
+			IgnoreConfig: &plugin.IgnoreConfig{
+				ShouldIgnoreErrorFunc: isIgnorableErrorPredicate([]string{"Request_ResourceNotFound", "Invalid object identifier"}),
+			},
+			KeyColumns: plugin.SingleColumn("id"),
 		},
 		List: &plugin.ListConfig{
-			Hydrate:           listAdConditionalAccessPolicies,
-			ShouldIgnoreError: isNotFoundErrorPredicate([]string{"Request_UnsupportedQuery"}),
+			Hydrate: listAdConditionalAccessPolicies,
+			IgnoreConfig: &plugin.IgnoreConfig{
+				ShouldIgnoreErrorFunc: isIgnorableErrorPredicate([]string{"Request_UnsupportedQuery"}),
+			},
 			KeyColumns: []*plugin.KeyColumn{
 				{Name: "display_name", Require: plugin.Optional},
 				{Name: "state", Require: plugin.Optional},
@@ -33,32 +42,32 @@ func tableAzureAdConditionalAccessPolicy() *plugin.Table {
 		},
 
 		Columns: []*plugin.Column{
-			{Name: "id", Type: proto.ColumnType_STRING, Description: "Specifies the identifier of a conditionalAccessPolicy object.", Transform: transform.FromGo()},
-			{Name: "display_name", Type: proto.ColumnType_STRING, Description: "Specifies a display name for the conditionalAccessPolicy object."},
-			{Name: "state", Type: proto.ColumnType_STRING, Description: "Specifies the state of the conditionalAccessPolicy object. Possible values are: enabled, disabled, enabledForReportingButNotEnforced."},
-			{Name: "created_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The create date of the conditional access policy."},
-			{Name: "modified_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The modification date of the conditional access policy."},
-			{Name: "operator", Type: proto.ColumnType_STRING, Description: "Defines the relationship of the grant controls. Possible values: AND, OR."},
+			{Name: "id", Type: proto.ColumnType_STRING, Description: "Specifies the identifier of a conditionalAccessPolicy object.", Transform: transform.FromMethod("GetId")},
+			{Name: "display_name", Type: proto.ColumnType_STRING, Description: "Specifies a display name for the conditionalAccessPolicy object.", Transform: transform.FromMethod("GetDisplayName")},
+			{Name: "state", Type: proto.ColumnType_STRING, Description: "Specifies the state of the conditionalAccessPolicy object. Possible values are: enabled, disabled, enabledForReportingButNotEnforced.", Transform: transform.FromMethod("GetState")},
+			{Name: "created_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The create date of the conditional access policy.", Transform: transform.FromMethod("GetCreatedDateTime")},
+			{Name: "modified_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The modification date of the conditional access policy.", Transform: transform.FromMethod("GetModifiedDateTime")},
+			{Name: "operator", Type: proto.ColumnType_STRING, Description: "Defines the relationship of the grant controls. Possible values: AND, OR.", Transform: transform.FromMethod("ConditionalAccessPolicyGrantControlsOperator")},
 
 			// Json fields
-			{Name: "applications", Type: proto.ColumnType_JSON, Description: "Applications and user actions included in and excluded from the policy.", Transform: transform.FromField("Conditions.Applications")},
-			{Name: "application_enforced_restrictions", Type: proto.ColumnType_JSON, Description: "Session control to enforce application restrictions. Only Exchange Online and Sharepoint Online support this session control.", Transform: transform.FromField("SessionControls.ApplicationEnforcedRestrictions")},
-			{Name: "built_in_controls", Type: proto.ColumnType_JSON, Description: "List of values of built-in controls required by the policy. Possible values: block, mfa, compliantDevice, domainJoinedDevice, approvedApplication, compliantApplication, passwordChange, unknownFutureValue.", Transform: transform.FromField("GrantControls.BuiltInControls")},
-			{Name: "client_app_types", Type: proto.ColumnType_JSON, Description: "Client application types included in the policy. Possible values are: all, browser, mobileAppsAndDesktopClients, exchangeActiveSync, easSupported, other.", Transform: transform.FromField("Conditions.ClientAppTypes")},
-			{Name: "custom_authentication_factors", Type: proto.ColumnType_JSON, Description: "List of custom controls IDs required by the policy.", Transform: transform.FromField("GrantControls.CustomAuthenticationFactors")},
-			{Name: "cloud_app_security", Type: proto.ColumnType_JSON, Description: "Session control to apply cloud app security.", Transform: transform.FromField("SessionControls.CloudAppSecurity")},
-			{Name: "locations", Type: proto.ColumnType_JSON, Description: "Locations included in and excluded from the policy.", Transform: transform.FromField("Conditions.Locations")},
-			{Name: "persistent_browser", Type: proto.ColumnType_JSON, Description: "Session control to define whether to persist cookies or not. All apps should be selected for this session control to work correctly.", Transform: transform.FromField("SessionControls.PersistentBrowser")},
-			{Name: "platforms", Type: proto.ColumnType_JSON, Description: "Platforms included in and excluded from the policy.", Transform: transform.FromField("Conditions.Platforms")},
-			{Name: "sign_in_frequency", Type: proto.ColumnType_JSON, Description: "Session control to enforce signin frequency.", Transform: transform.FromField("SessionControls.SignInFrequency")},
-			{Name: "sign_in_risk_levels", Type: proto.ColumnType_JSON, Description: "Sign-in risk levels included in the policy. Possible values are: low, medium, high, hidden, none, unknownFutureValue.", Transform: transform.FromField("Conditions.SignInRiskLevels")},
-			{Name: "terms_of_use", Type: proto.ColumnType_JSON, Description: "List of terms of use IDs required by the policy.", Transform: transform.FromField("GrantControls.TermsOfUse")},
-			{Name: "users", Type: proto.ColumnType_JSON, Description: "Users, groups, and roles included in and excluded from the policy.", Transform: transform.FromField("Conditions.Users")},
-			{Name: "user_risk_levels", Type: proto.ColumnType_JSON, Description: "User risk levels included in the policy. Possible values are: low, medium, high, hidden, none, unknownFutureValue.", Transform: transform.FromField("Conditions.UserRiskLevels")},
+			{Name: "applications", Type: proto.ColumnType_JSON, Description: "Applications and user actions included in and excluded from the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsApplications")},
+			{Name: "application_enforced_restrictions", Type: proto.ColumnType_JSON, Description: "Session control to enforce application restrictions. Only Exchange Online and Sharepoint Online support this session control.", Transform: transform.FromMethod("ConditionalAccessPolicySessionControlsApplicationEnforcedRestrictions")},
+			{Name: "built_in_controls", Type: proto.ColumnType_JSON, Description: "List of values of built-in controls required by the policy. Possible values: block, mfa, compliantDevice, domainJoinedDevice, approvedApplication, compliantApplication, passwordChange, unknownFutureValue.", Transform: transform.FromMethod("ConditionalAccessPolicyGrantControlsBuiltInControls")},
+			{Name: "client_app_types", Type: proto.ColumnType_JSON, Description: "Client application types included in the policy. Possible values are: all, browser, mobileAppsAndDesktopClients, exchangeActiveSync, easSupported, other.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsClientAppTypes")},
+			{Name: "custom_authentication_factors", Type: proto.ColumnType_JSON, Description: "List of custom controls IDs required by the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyGrantControlsCustomAuthenticationFactors")},
+			{Name: "cloud_app_security", Type: proto.ColumnType_JSON, Description: "Session control to apply cloud app security.", Transform: transform.FromMethod("ConditionalAccessPolicySessionControlsCloudAppSecurity")},
+			{Name: "locations", Type: proto.ColumnType_JSON, Description: "Locations included in and excluded from the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsLocations")},
+			{Name: "persistent_browser", Type: proto.ColumnType_JSON, Description: "Session control to define whether to persist cookies or not. All apps should be selected for this session control to work correctly.", Transform: transform.FromMethod("ConditionalAccessPolicySessionControlsPersistentBrowser")},
+			{Name: "platforms", Type: proto.ColumnType_JSON, Description: "Platforms included in and excluded from the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsPlatforms")},
+			{Name: "sign_in_frequency", Type: proto.ColumnType_JSON, Description: "Session control to enforce signin frequency.", Transform: transform.FromMethod("ConditionalAccessPolicySessionControlsSignInFrequency")},
+			{Name: "sign_in_risk_levels", Type: proto.ColumnType_JSON, Description: "Sign-in risk levels included in the policy. Possible values are: low, medium, high, hidden, none, unknownFutureValue.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsSignInRiskLevels")},
+			{Name: "terms_of_use", Type: proto.ColumnType_JSON, Description: "List of terms of use IDs required by the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyGrantControlsTermsOfUse")},
+			{Name: "users", Type: proto.ColumnType_JSON, Description: "Users, groups, and roles included in and excluded from the policy.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsUsers")},
+			{Name: "user_risk_levels", Type: proto.ColumnType_JSON, Description: "User risk levels included in the policy. Possible values are: low, medium, high, hidden, none, unknownFutureValue.", Transform: transform.FromMethod("ConditionalAccessPolicyConditionsUserRiskLevels")},
 
 			// Standard columns
-			{Name: "title", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTitle, Transform: transform.FromField("DisplayName", "ID")},
-			{Name: "tenant_id", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTenant, Hydrate: plugin.HydrateFunc(getTenantId).WithCache(), Transform: transform.FromValue()},
+			{Name: "title", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTitle, Transform: transform.From(adConditionalAccessPolicyTitle)},
+			{Name: "tenant_id", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTenant, Hydrate: plugin.HydrateFunc(getTenant).WithCache(), Transform: transform.FromValue()},
 		},
 	}
 }
@@ -66,88 +75,116 @@ func tableAzureAdConditionalAccessPolicy() *plugin.Table {
 //// LIST FUNCTION
 
 func listAdConditionalAccessPolicies(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
-	session, err := GetNewSession(ctx, d)
+	// Create client
+	client, adapter, err := GetGraphClient(ctx, d)
 	if err != nil {
-		return nil, err
+		return nil, errors.New(fmt.Sprintf("error creating client: %v", err))
 	}
 
-	client := msgraph.NewConditionalAccessPolicyClient(session.TenantID)
-	client.BaseClient.Authorizer = session.Authorizer
-
-	input := odata.Query{}
+	// List operations
+	input := &policies.PoliciesRequestBuilderGetQueryParameters{}
 
 	// Restrict the limit value to be passed in the query parameter which is not between 1 and 999, otherwise API will throw an error as follow
 	// unexpected status 400 with OData error: Request_UnsupportedQuery: Invalid page size specified: '1000'. Must be between 1 and 999 inclusive.
 	limit := d.QueryContext.Limit
 	if limit != nil {
 		if *limit > 0 && *limit <= 999 {
-			input.Top = int(*limit)
+			l := int32(*limit)
+			input.Top = &l
 		}
 	}
-	
-	qualsColumnMap := []QualsColumn{
-		{"display_name", "string", "displayName"},
-		{"state", "string", "state"},
-	}
 
-	filter := buildCommaonQueryFilter(qualsColumnMap, d.Quals)
+	equalQuals := d.KeyColumnQuals
+	filter := buildConditionalAccessPolicyQueryFilter(equalQuals)
+
 	if len(filter) > 0 {
-		input.Filter = strings.Join(filter, " and ")
+		joinStr := strings.Join(filter, " and ")
+		input.Filter = &joinStr
 	}
 
-	conditionalAccessPolicies, _, err := client.List(ctx, input)
+	options := &policies.PoliciesRequestBuilderGetRequestConfiguration{
+		QueryParameters: input,
+	}
+
+	result, err := client.Identity().ConditionalAccess().Policies().GetWithRequestConfigurationAndResponseHandler(options, nil)
 	if err != nil {
-		if isNotFoundError(err) {
-			plugin.Logger(ctx).Error("listAdConditionalAccessPolicies", "Resource not found", err)
-			return nil, nil
-		}
-		return nil, err
+		errObj := getErrorObject(err)
+		return nil, errObj
 	}
 
-	for _, conditionalAccesPolicy := range *conditionalAccessPolicies {
-		d.StreamListItem(ctx, conditionalAccesPolicy)
+	pageIterator, err := msgraphcore.NewPageIterator(result, adapter, models.CreateConditionalAccessPolicyCollectionResponseFromDiscriminatorValue)
+
+	err = pageIterator.Iterate(func(pageItem interface{}) bool {
+		policy := pageItem.(models.ConditionalAccessPolicyable)
+
+		d.StreamListItem(ctx, &ADConditionalAccessPolicyInfo{policy})
 
 		// Context can be cancelled due to manual cancellation or the limit has been hit
 		if d.QueryStatus.RowsRemaining(ctx) == 0 {
-			return nil, nil
+			return false
 		}
-	}
+
+		return true
+	})
+
 	return nil, err
 }
 
-//// Hydrate Functions
+//// HYDRATE FUNCTIONS
 
 func getAdConditionalAccessPolicy(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	var conditionalAccessPolicyId string
-	if h.Item != nil {
-		conditionalAccessPolicyId = *h.Item.(msgraph.ConditionalAccessPolicy).ID
-	} else {
-		conditionalAccessPolicyId = d.KeyColumnQuals["id"].GetStringValue()
-	}
 
+	conditionalAccessPolicyId := d.KeyColumnQuals["id"].GetStringValue()
 	if conditionalAccessPolicyId == "" {
 		return nil, nil
 	}
 
-	session, err := GetNewSession(ctx, d)
+	// Create client
+	client, _, err := GetGraphClient(ctx, d)
 	if err != nil {
-		if isNotFoundError(err) {
-			return nil, nil
-		}
-		return nil, err
+		return nil, errors.New(fmt.Sprintf("error creating client: %v", err))
 	}
 
-	client := msgraph.NewConditionalAccessPolicyClient(session.TenantID)
-	client.BaseClient.Authorizer = session.Authorizer
-	client.BaseClient.DisableRetries = true
-
-	conditionalAccessPolicy, _, err := client.Get(ctx, conditionalAccessPolicyId, odata.Query{})
+	policy, err := client.Identity().ConditionalAccess().PoliciesById(conditionalAccessPolicyId).Get()
 	if err != nil {
-		if isNotFoundError(err) {
-			plugin.Logger(ctx).Error("getAdConditionalAccessPolicy", "Resource not found", err)
-			return nil, nil
-		}
-		return nil, err
+		errObj := getErrorObject(err)
+		return nil, errObj
 	}
-	return conditionalAccessPolicy, nil
+	return &ADConditionalAccessPolicyInfo{policy}, nil
+}
+
+func buildConditionalAccessPolicyQueryFilter(equalQuals plugin.KeyColumnEqualsQualMap) []string {
+	filters := []string{}
+
+	filterQuals := map[string]string{
+		"display_name": "string",
+		"state":        "string",
+	}
+
+	for qual, qualType := range filterQuals {
+		switch qualType {
+		case "string":
+			if equalQuals[qual] != nil {
+				filters = append(filters, fmt.Sprintf("%s eq '%s'", strcase.ToCamel(qual), equalQuals[qual].GetStringValue()))
+			}
+		}
+	}
+
+	return filters
+}
+
+//// TRANSFORM FUNCTIONS
+
+func adConditionalAccessPolicyTitle(_ context.Context, d *transform.TransformData) (interface{}, error) {
+	data := d.HydrateItem.(*ADConditionalAccessPolicyInfo)
+	if data == nil {
+		return nil, nil
+	}
+
+	title := data.GetDisplayName()
+	if title == nil {
+		title = data.GetId()
+	}
+
+	return title, nil
 }
