+++ azuread/table_azuread_user.go
@@ -2,17 +2,20 @@ package azuread
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"os"
 	"strings"
 
-	"github.com/ettle/strcase"
-	"github.com/manicminer/hamilton/msgraph"
-	"github.com/manicminer/hamilton/odata"
-	"github.com/turbot/go-kit/helpers"
-	"github.com/turbot/steampipe-plugin-sdk/v3/grpc/proto"
-	"github.com/turbot/steampipe-plugin-sdk/v3/plugin/transform"
+	"github.com/iancoleman/strcase"
+	msgraphcore "github.com/microsoftgraph/msgraph-sdk-go-core"
+	"github.com/microsoftgraph/msgraph-sdk-go/models"
+	"github.com/microsoftgraph/msgraph-sdk-go/users"
+	"github.com/microsoftgraph/msgraph-sdk-go/users/item"
 
+	"github.com/turbot/steampipe-plugin-sdk/v3/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v3/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v3/plugin/transform"
 )
 
 //// TABLE DEFINITION
@@ -22,15 +25,16 @@ func tableAzureAdUser() *plugin.Table {
 		Name:        "azuread_user",
 		Description: "Represents an Azure AD user account.",
 		Get: &plugin.GetConfig{
-			Hydrate:           getAdUser,
-			ShouldIgnoreError: isNotFoundErrorPredicate([]string{"Invalid object identifier"}),
-			KeyColumns:        plugin.SingleColumn("id"),
+			Hydrate: getAdUser,
+			IgnoreConfig: &plugin.IgnoreConfig{
+				ShouldIgnoreErrorFunc: isIgnorableErrorPredicate([]string{"Request_ResourceNotFound", "Invalid object identifier"}),
+			},
+			KeyColumns: plugin.SingleColumn("id"),
 		},
 		List: &plugin.ListConfig{
 			Hydrate: listAdUsers,
 			KeyColumns: plugin.KeyColumnSlice{
 				// Key fields
-				{Name: "id", Require: plugin.Optional},
 				{Name: "user_principal_name", Require: plugin.Optional},
 				{Name: "filter", Require: plugin.Optional},
 
@@ -43,36 +47,35 @@ func tableAzureAdUser() *plugin.Table {
 		},
 
 		Columns: []*plugin.Column{
-			{Name: "display_name", Type: proto.ColumnType_STRING, Description: "The name displayed in the address book for the user. This is usually the combination of the user's first name, middle initial and last name."},
-			{Name: "id", Type: proto.ColumnType_STRING, Description: "The unique identifier for the user. Should be treated as an opaque identifier.", Transform: transform.FromGo()},
-			{Name: "user_principal_name", Type: proto.ColumnType_STRING, Description: "Principal email of the active directory user."},
-			{Name: "account_enabled", Type: proto.ColumnType_BOOL, Description: "True if the account is enabled; otherwise, false."},
-			{Name: "user_type", Type: proto.ColumnType_STRING, Description: "A string value that can be used to classify user types in your directory."},
-			{Name: "given_name", Type: proto.ColumnType_STRING, Description: "The given name (first name) of the user."},
-			{Name: "surname", Type: proto.ColumnType_STRING, Description: "Family name or last name of the active directory user."},
+			{Name: "display_name", Type: proto.ColumnType_STRING, Description: "The name displayed in the address book for the user. This is usually the combination of the user's first name, middle initial and last name.", Transform: transform.FromMethod("GetDisplayName")},
+			{Name: "id", Type: proto.ColumnType_STRING, Description: "The unique identifier for the user. Should be treated as an opaque identifier.", Transform: transform.FromMethod("GetId")},
+			{Name: "user_principal_name", Type: proto.ColumnType_STRING, Description: "Principal email of the active directory user.", Transform: transform.FromMethod("GetUserPrincipalName")},
+			{Name: "account_enabled", Type: proto.ColumnType_BOOL, Description: "True if the account is enabled; otherwise, false.", Transform: transform.FromMethod("GetAccountEnabled")},
+			{Name: "user_type", Type: proto.ColumnType_STRING, Description: "A string value that can be used to classify user types in your directory.", Transform: transform.FromMethod("GetUserType")},
+			{Name: "given_name", Type: proto.ColumnType_STRING, Description: "The given name (first name) of the user.", Transform: transform.FromMethod("GetGivenName")},
+			{Name: "surname", Type: proto.ColumnType_STRING, Description: "Family name or last name of the active directory user.", Transform: transform.FromMethod("GetSurname")},
 
 			{Name: "filter", Type: proto.ColumnType_STRING, Transform: transform.FromQual("filter"), Description: "Odata query to search for resources."},
 
 			// Other fields
-			{Name: "on_premises_immutable_id", Type: proto.ColumnType_STRING, Description: "Used to associate an on-premises Active Directory user account with their Azure AD user object."},
-			{Name: "created_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The time at which the user was created."},
-			{Name: "mail", Type: proto.ColumnType_STRING, Description: "The SMTP address for the user, for example, jeff@contoso.onmicrosoft.com."},
-			{Name: "mail_nickname", Type: proto.ColumnType_STRING, Description: "The mail alias for the user."},
-			{Name: "password_policies", Type: proto.ColumnType_STRING, Description: "Specifies password policies for the user. This value is an enumeration with one possible value being DisableStrongPassword, which allows weaker passwords than the default policy to be specified. DisablePasswordExpiration can also be specified. The two may be specified together; for example: DisablePasswordExpiration, DisableStrongPassword."},
-			{Name: "refresh_tokens_valid_from_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "Any refresh tokens or sessions tokens (session cookies) issued before this time are invalid, and applications will get an error when using an invalid refresh or sessions token to acquire a delegated access token (to access APIs such as Microsoft Graph)."},
-			{Name: "sign_in_sessions_valid_from_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "Any refresh tokens or sessions tokens (session cookies) issued before this time are invalid, and applications will get an error when using an invalid refresh or sessions token to acquire a delegated access token (to access APIs such as Microsoft Graph)."},
-			{Name: "usage_location", Type: proto.ColumnType_STRING, Description: "A two letter country code (ISO standard 3166), required for users that will be assigned licenses due to legal requirement to check for availability of services in countries."},
+			{Name: "on_premises_immutable_id", Type: proto.ColumnType_STRING, Description: "Used to associate an on-premises Active Directory user account with their Azure AD user object.", Transform: transform.FromMethod("GetOnPremisesImmutableId")},
+			{Name: "created_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "The time at which the user was created.", Transform: transform.FromMethod("GetCreatedDateTime")},
+			{Name: "mail", Type: proto.ColumnType_STRING, Description: "The SMTP address for the user, for example, jeff@contoso.onmicrosoft.com.", Transform: transform.FromMethod("GetMail")},
+			{Name: "mail_nickname", Type: proto.ColumnType_STRING, Description: "The mail alias for the user.", Transform: transform.FromMethod("GetMailNickname")},
+			{Name: "password_policies", Type: proto.ColumnType_STRING, Description: "Specifies password policies for the user. This value is an enumeration with one possible value being DisableStrongPassword, which allows weaker passwords than the default policy to be specified. DisablePasswordExpiration can also be specified. The two may be specified together; for example: DisablePasswordExpiration, DisableStrongPassword.", Transform: transform.FromMethod("GetPasswordPolicies")},
+			// {Name: "refresh_tokens_valid_from_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "Any refresh tokens or sessions tokens (session cookies) issued before this time are invalid, and applications will get an error when using an invalid refresh or sessions token to acquire a delegated access token (to access APIs such as Microsoft Graph)."},
+			{Name: "sign_in_sessions_valid_from_date_time", Type: proto.ColumnType_TIMESTAMP, Description: "Any refresh tokens or sessions tokens (session cookies) issued before this time are invalid, and applications will get an error when using an invalid refresh or sessions token to acquire a delegated access token (to access APIs such as Microsoft Graph).", Transform: transform.FromMethod("GetSignInSessionsValidFromDateTime")},
+			{Name: "usage_location", Type: proto.ColumnType_STRING, Description: "A two letter country code (ISO standard 3166), required for users that will be assigned licenses due to legal requirement to check for availability of services in countries.", Transform: transform.FromMethod("GetUsageLocation")},
 
 			// Json fields
-			{Name: "member_of", Type: proto.ColumnType_JSON, Description: "A list the groups and directory roles that the user is a direct member of."},
-			{Name: "additional_properties", Type: proto.ColumnType_JSON, Description: "A list of unmatched properties from the message are deserialized this collection."},
-			{Name: "im_addresses", Type: proto.ColumnType_JSON, Description: "The instant message voice over IP (VOIP) session initiation protocol (SIP) addresses for the user."},
-			{Name: "other_mails", Type: proto.ColumnType_JSON, Description: "A list of additional email addresses for the user."},
-			{Name: "password_profile", Type: proto.ColumnType_JSON, Description: "Specifies the password profile for the user. The profile contains the user’s password. This property is required when a user is created."},
+			{Name: "member_of", Type: proto.ColumnType_JSON, Description: "A list the groups and directory roles that the user is a direct member of.", Transform: transform.FromMethod("UserMemberOf")},
+			{Name: "im_addresses", Type: proto.ColumnType_JSON, Description: "The instant message voice over IP (VOIP) session initiation protocol (SIP) addresses for the user.", Transform: transform.FromMethod("GetImAddresses")},
+			{Name: "other_mails", Type: proto.ColumnType_JSON, Description: "A list of additional email addresses for the user.", Transform: transform.FromMethod("GetOtherMails")},
+			{Name: "password_profile", Type: proto.ColumnType_JSON, Description: "Specifies the password profile for the user. The profile contains the user’s password. This property is required when a user is created.", Transform: transform.FromMethod("UserPasswordProfile")},
 
 			// Standard columns
-			{Name: "title", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTitle, Transform: transform.FromField("DisplayName", "UserPrincipalName")},
-			{Name: "tenant_id", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTenant, Hydrate: plugin.HydrateFunc(getTenantId).WithCache(), Transform: transform.FromValue()},
+			{Name: "title", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTitle, Transform: transform.From(adUserTitle)},
+			{Name: "tenant_id", Type: proto.ColumnType_STRING, Description: ColumnDescriptionTenant, Hydrate: plugin.HydrateFunc(getTenant).WithCache(), Transform: transform.FromValue()},
 		},
 	}
 }
@@ -80,34 +83,33 @@ func tableAzureAdUser() *plugin.Table {
 //// LIST FUNCTION
 
 func listAdUsers(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData) (interface{}, error) {
-	session, err := GetNewSession(ctx, d)
+	// Create client
+	client, adapter, err := GetGraphClient(ctx, d)
 	if err != nil {
-		return nil, err
+		return nil, errors.New(fmt.Sprintf("error creating client: %v", err))
 	}
 
-	client := msgraph.NewUsersClient(session.TenantID)
-	client.BaseClient.Authorizer = session.Authorizer
+	// List operations
+	input := &users.UsersRequestBuilderGetQueryParameters{}
 
-	input := odata.Query{}
-	
 	// Restrict the limit value to be passed in the query parameter which is not between 1 and 999, otherwise API will throw an error as follow
-	// unexpected status 400 with OData error: Request_UnsupportedQuery: Invalid page size specified: '1000'. Must be between 1 and 999 inclusive.
 	limit := d.QueryContext.Limit
 	if limit != nil {
 		if *limit > 0 && *limit <= 999 {
-			input.Top = int(*limit)
-		}
-	}
-
-	if helpers.StringSliceContains(d.QueryContext.Columns, "member_of") {
-		input.Expand = odata.Expand{
-			Relationship: "memberOf",
-			Select:       []string{"id", "displayName"},
+			l := int32(*limit)
+			input.Top = &l
 		}
 	}
 
 	equalQuals := d.KeyColumnQuals
 	quals := d.Quals
+
+	// Check for query context and requests only for queried columns
+	givenColumns := d.QueryContext.Columns
+	selectColumns, expandColumns := buildUserRequestFields(ctx, givenColumns)
+
+	input.Select = selectColumns
+	input.Expand = expandColumns
 
 	var queryFilter string
 	filter := buildQueryFilter(equalQuals)
@@ -118,78 +120,137 @@ func listAdUsers(ctx context.Context, d *plugin.QueryData, _ *plugin.HydrateData
 	}
 
 	if queryFilter != "" {
-		input.Filter = queryFilter
+		input.Filter = &queryFilter
 	} else if len(filter) > 0 {
-		input.Filter = strings.Join(filter, " and ")
+		joinStr := strings.Join(filter, " and ")
+		input.Filter = &joinStr
 	}
 
-	users, _, err := client.List(ctx, input)
+	options := &users.UsersRequestBuilderGetRequestConfiguration{
+		QueryParameters: input,
+	}
+
+	result, err := client.Users().GetWithRequestConfigurationAndResponseHandler(options, nil)
 	if err != nil {
-		if isNotFoundError(err) {
-			return nil, nil
-		}
-		return nil, err
+		errObj := getErrorObject(err)
+		return nil, errObj
 	}
 
-	for _, user := range *users {
-		d.StreamListItem(ctx, user)
+	pageIterator, err := msgraphcore.NewPageIterator(result, adapter, models.CreateUserCollectionResponseFromDiscriminatorValue)
+
+	err = pageIterator.Iterate(func(pageItem interface{}) bool {
+		user := pageItem.(models.Userable)
+
+		d.StreamListItem(ctx, &ADUserInfo{user})
 
 		// Context can be cancelled due to manual cancellation or the limit has been hit
 		if d.QueryStatus.RowsRemaining(ctx) == 0 {
-			return nil, nil
+			return false
 		}
-	}
 
-	return nil, err
+		return true
+	})
+
+	return nil, nil
 }
 
 //// HYDRATE FUNCTIONS
 
 func getAdUser(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
 
-	var userId string
-	if h.Item != nil {
-		userId = *h.Item.(msgraph.User).ID
-	} else {
-		userId = d.KeyColumnQuals["id"].GetStringValue()
+	// Create client
+	client, _, err := GetGraphClient(ctx, d)
+	if err != nil {
+		return nil, errors.New(fmt.Sprintf("error creating client: %v", err))
 	}
 
+	userId := d.KeyColumnQuals["id"].GetStringValue()
 	if userId == "" {
 		return nil, nil
 	}
-	session, err := GetNewSession(ctx, d)
-	if err != nil {
-		return nil, err
+
+	// Check for query context and requests only for queried columns
+	givenColumns := d.QueryContext.Columns
+	selectColumns, expandColumns := buildUserRequestFields(ctx, givenColumns)
+
+	input := &item.UserItemRequestBuilderGetQueryParameters{}
+	input.Select = selectColumns
+	input.Expand = expandColumns
+
+	options := &item.UserItemRequestBuilderGetRequestConfiguration{
+		QueryParameters: input,
 	}
 
-	client := msgraph.NewUsersClient(session.TenantID)
-	client.BaseClient.Authorizer = session.Authorizer
-	client.BaseClient.DisableRetries = true
+	user, err := client.UsersById(userId).GetWithRequestConfigurationAndResponseHandler(options, nil)
+	if err != nil {
+		errObj := getErrorObject(err)
+		return nil, errObj
+	}
 
-	input := odata.Query{}
-	if helpers.StringSliceContains(d.QueryContext.Columns, "member_of") {
-		input.Expand = odata.Expand{
-			Relationship: "memberOf",
-			Select:       []string{"id", "displayName"},
+	return &ADUserInfo{user}, nil
+}
+
+func buildUserRequestFields(ctx context.Context, queryColumns []string) ([]string, []string) {
+	var selectColumns, expandColumns []string
+
+	for _, columnName := range queryColumns {
+		if columnName == "title" || columnName == "filter" || columnName == "tenant_id" {
+			continue
+		}
+
+		if columnName == "member_of" {
+			expandColumns = append(expandColumns, fmt.Sprintf("%s($select=id,displayName)", strcase.ToLowerCamel(columnName)))
+			continue
+		}
+
+		if columnName == "refresh_tokens_valid_from_date_time" {
+			selectColumns = append(selectColumns, "signInSessionsValidFromDateTime")
+			continue
+		}
+
+		selectColumns = append(selectColumns, strcase.ToLowerCamel(columnName))
+	}
+
+	return selectColumns, expandColumns
+}
+
+func getTenant(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
+	var tenantID string
+	var err error
+
+	// Read tenantID from config, or environment variables
+	azureADConfig := GetConfig(d.Connection)
+	if azureADConfig.TenantID != nil {
+		tenantID = *azureADConfig.TenantID
+	} else {
+		tenantID = os.Getenv("AZURE_TENANT_ID")
+	}
+
+	// If not set in config, get tenantID from CLI
+	if tenantID == "" {
+		tenantID, err = getTenantFromCLI()
+		if err != nil {
+			return nil, err
 		}
 	}
 
-	user, _, err := client.Get(ctx, userId, input)
-	if err != nil {
-		return nil, err
-	}
-	return *user, nil
+	return tenantID, nil
 }
 
-func getTenantId(ctx context.Context, d *plugin.QueryData, h *plugin.HydrateData) (interface{}, error) {
-	plugin.Logger(ctx).Debug("getTenantId")
+//// TRANSFORM FUNCTIONS
 
-	session, err := GetNewSession(ctx, d)
-	if err != nil {
-		return nil, err
+func adUserTitle(_ context.Context, d *transform.TransformData) (interface{}, error) {
+	data := d.HydrateItem.(*ADUserInfo)
+	if data == nil {
+		return nil, nil
 	}
 
-	return session.TenantID, nil
+	title := data.GetDisplayName()
+	if title == nil {
+		title = data.GetUserPrincipalName()
+	}
+
+	return title, nil
 }
 
 func buildQueryFilter(equalQuals plugin.KeyColumnEqualsQualMap) []string {
